feat: location & coordinate saved to localstorage logic added
