+++ src/app/setting/location/page.tsx
@@ -4,10 +4,11 @@ import Button from "@/componants/button/Button";
 import Header from "@/componants/header/Header";
 import Map from "@/componants/map/Map";
 import SearchBar from "@/componants/searchBar/SearchBar";
-import { useGetAddress } from "@/hooks/useGetAddress";
+import { useHandleAddress } from "@/hooks/useHandleAddress";
 
 export default function Location() {
-  const { handleClick, address } = useGetAddress();
+  const { handleClick, handleSubmit, handleCoordinate, coordinate, address } =
+    useHandleAddress();
 
   return (
     <>
@@ -15,9 +16,13 @@ export default function Location() {
       <section className="grid place-items-center gap-10">
         <div className="grid grid-cols-1 gap-12 w-full place-items-center my-10">
           <SearchBar onClick={handleClick} address={address} />
-          <Map address={address} />
+          <Map
+            address={address}
+            coordinate={coordinate}
+            handleCoordinate={handleCoordinate}
+          />
         </div>
-        <Button>저장하기</Button>
+        <Button onClick={handleSubmit}>저장하기</Button>
       </section>
     </>
   );
