+++ src/componants/map/Map.tsx
@@ -4,34 +4,13 @@ import { Map as KakaoMap, Loader, MapMarker } from "react-kakao-maps-sdk";
 
 interface MapProps {
   address: string;
+  coordinate: { lat: number; lng: number };
+  handleCoordinate: () => void;
 }
 
-const Map = ({ address }: MapProps) => {
-  const [coordinate, setCoordinate] = useState(INITIAL_COORDINATE_DATA);
-
+const Map = ({ address, handleCoordinate, coordinate }: MapProps) => {
   useEffect(() => {
-    new Loader({
-      appkey: `${process.env.NEXT_PUBLIC_KAKAO_APP_KEY}`,
-      libraries: ["services"],
-    })
-      .load()
-      .then((kakao) => {
-        if (address.length > 0) {
-          const geocoder = new kakao.maps.services.Geocoder();
-          geocoder.addressSearch(address, function (result, status) {
-            if (status === kakao.maps.services.Status.OK) {
-              setCoordinate({
-                lat: parseFloat(result[0].y),
-                lng: parseFloat(result[0].x),
-              });
-            } else {
-              alert(
-                "해당하는 주소를 지도에서 찾지 못했습니다.\n잠시후 다시 시도해주세요"
-              );
-            }
-          });
-        }
-      });
+    handleCoordinate();
   }, [address]);
 
   return (
