+++ src/constants/location.ts
@@ -1,4 +1,6 @@
 export const INITIAL_COORDINATE_DATA = {
-  lat: 33.450701,
-  lng: 126.570667,
+  lat: 37.556103,
+  lng: 126.971551,
 };
+
+export const INITIAL_LOCATION = "서울특별시 중구 통일로 1";
