--- src/hooks/useGetAddress.tsx
@@ -1,30 +1,0 @@
-import { useState } from "react";
-import { Address, useDaumPostcodePopup } from "react-daum-postcode";
-
-export const useGetAddress = () => {
-  const [address, setAddress] = useState("");
-  const open = useDaumPostcodePopup();
-
-  const handleComplete = (data: Address) => {
-    let fullAddress = data.address;
-    let extraAddress = "";
-
-    if (data.addressType === "R") {
-      if (data.bname !== "") {
-        extraAddress += data.bname;
-      }
-      if (data.buildingName !== "") {
-        extraAddress +=
-          extraAddress !== "" ? `, ${data.buildingName}` : data.buildingName;
-      }
-      fullAddress += extraAddress !== "" ? ` (${extraAddress})` : "";
-    }
-    setAddress(fullAddress);
-  };
-
-  const handleClick = () => {
-    open({ onComplete: handleComplete });
-  };
-
-  return { handleClick, address };
-};
